DENSO-main/requirement.txt
opencv-python 
numpy 
torch
ultralytics
pyserial
math